@@ -10,7 +10,7 @@
 		"start:server:dev": "node server/server.js serve debug",
 		"build": "react-scripts build",
 		"test": "yarn test:unit && yarn test:e2e",
-		"test:unit": "CI=true react-scripts test",
+		"test:unit": "jest --config jest.config.js",
 		"test:unit:watch": "react-scripts test",
 		"test:unit:cover": "CI=true react-scripts test --coverage",
 		"test:e2e": "PORT=3000 start-server-and-test test:e2e:server 3000 test:e2e:run",
@@ -25,39 +25,31 @@
 		"nuke": "rm -rf build && rm -rf node_modules && rm yarn.lock"
 	},
 	"dependencies": {
-		"@emotion/react": "^11.11.1",
-		"@reach/router": "^1.3.4",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-spring": "^9.7.2",
-		"use-sound": "^4.0.1"
+		"@emotion/react": "^11.14.0",
+		"react": "^19.0.0",
+		"react-dom": "^19.0.0",
+		"react-router-dom": "^7.1.3",
+		"react-spring": "^9.7.5",
+		"use-sound": "^4.0.3"
 	},
 	"devDependencies": {
+		"@babel/core": "^7.26.7",
 		"@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-		"@testing-library/cypress": "^9.0.0",
-		"concurrently": "^8.2.0",
-		"cypress": "^12.17.1",
+		"@babel/preset-env": "^7.26.7",
+		"@babel/preset-react": "^7.26.3",
+		"@cypress/webpack-preprocessor": "^6.0.2",
+		"@testing-library/cypress": "^10.0.3",
+		"babel-jest": "^29.7.0",
+		"concurrently": "^9.1.2",
+		"cypress": "^14.0.1",
 		"cypress-waitfor": "^1.1.0",
+		"jest": "^29.7.0",
+		"jest-environment-jsdom": "^29.7.0",
+		"raw-loader": "^4.0.2",
 		"react-scripts": "5.0.1",
-		"start-server-and-test": "^2.0.0"
+		"start-server-and-test": "^2.0.10"
 	},
 	"proxy": "http://localhost:5555",
-	"jest": {
-		"collectCoverageFrom": [
-			"src/**/*.js",
-			"server/**/*.js",
-			"!<rootDir>/node_modules/",
-			"!<rootDir>/server/node_modules/"
-		],
-		"coverageThreshold": {
-			"global": {
-				"branches": 0,
-				"functions": 0,
-				"lines": 0,
-				"statements": 0
-			}
-		}
-	},
 	"browserslist": {
 		"production": [
 			">0.2%",
