@@ -1,8 +1,11 @@
 import * as serviceWorker from './serviceWorker';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import React from 'react';
 import App from './app';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(<App />);
 
 serviceWorker.register();
